backend/requirements.txt
fastapi
uvicorn
pandas
google-genai
python-dotenv
pydantic